create diagnostics directory
