+++ distributed/diagnostics/__init__.py
@@ -1,0 +1,1 @@
+from .progress import progress
